+++ .flake8
@@ -1,3 +1,3 @@
 [flake8]
 max-line-length=119
-exclude=migrations, .git, build.sh, .yml
+exclude=migrations, .git, build.sh, .yml, __init__.py
