+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
 env =
-    CLIENT_ID=1
-    CLIENT_SECRET=F
-    USER_ID=3
+    CLIENT_ID = FAKE_CLIENT_ID
+    CLIENT_SECRET = FAKE_CLIENT_SECRET
+    USER_ID = FAKE_USER_ID
