+++ tests/vcr_cassettes/test_get_user_feeds.yaml
@@ -13,8 +13,8 @@ interactions:
     uri: https://api.mercadolibre.com/oauth/token
   response:
     body: {string: '{"access_token": "HIDDEN", "token_type": "bearer", "expires_in":
-        21600, "scope": "marketplace_off offline_access read reporting_full wallet-payments
-        write", "user_id": "HIDDEN", "refresh_token": "HIDDEN"}'}
+        21600, "scope": "offline_access read write", "user_id": "HIDDEN", "refresh_token":
+        "HIDDEN"}'}
     headers:
       Access-Control-Allow-Headers: [Content-Type]
       Access-Control-Allow-Methods: ['PUT, GET, POST, DELETE, OPTIONS']
@@ -23,17 +23,17 @@ interactions:
       Cache-Control: [max-age=0]
       Connection: [keep-alive]
       Content-Type: [application/json;charset=UTF-8]
-      Date: ['Thu, 20 Dec 2018 18:51:25 GMT']
-      ETag: [89851469f82e03aa50df0d5b83a2e0e0]
+      Date: ['Fri, 21 Dec 2018 07:45:06 GMT']
+      ETag: [de0251160f5bec09228cc72fe579a67f]
       Strict-Transport-Security: [max-age=15724800]
       Transfer-Encoding: [chunked]
       Vary: ['Accept,Accept-Encoding']
-      Via: [1.1 c3c79d4b701a0d344a641862e9711512.cloudfront.net (CloudFront)]
-      X-Amz-Cf-Id: [-w08f77dTlCeieTP9oiJW-n9tETID-iDraHOhuSlM45It6iKmBl2Jw==]
+      Via: [1.1 af3ec872aa77de6760622a554164a644.cloudfront.net (CloudFront)]
+      X-Amz-Cf-Id: [lpQhABHSLHsSHQ8YvZjP8hPvrK-RWSfBIlrXfVOs2h_UWFZ0Uov5pA==]
       X-Cache: [Miss from cloudfront]
       X-Content-Type-Options: [nosniff]
       X-Frame-Options: [DENY]
-      X-Request-Id: [4246cd7c-5ccf-43e3-aac2-7a9809a0830d]
+      X-Request-Id: [aa4065ce-24d9-4993-9be7-bf0afca977a7]
       X-XSS-Protection: [1; mode=block]
     status: {code: 200, message: OK}
 - request:
@@ -46,89 +46,37 @@ interactions:
     method: GET
     uri: https://api.mercadolibre.com/myfeeds
   response:
-    body: {string: '{"messages": []}'}
+    body: {string: '{"messages": [{"_id": "5c181060d42441b1716c9adb", "resource":
+        "/questions/5995837548", "user_id": "128828674", "topic": "questions", "application_id":
+        4607298613855815, "attempts": 1, "sent": "2018-12-17T21:08:48.586Z", "received":
+        "2018-12-17T21:08:48.527Z", "request": {"url": "https://localhost:8000/auth",
+        "headers": {"host": "localhost:8000", "accept": "application/json", "content-type":
+        "application/json", "content-length": 197}, "data": "{\"resource\":\"/questions/5995837548\",\"user_id\":\"128828674\",\"topic\":\"questions\",\"application_id\":4607298613855815,\"attempts\":1,\"sent\":\"2018-12-17T21:08:48.586Z\",\"received\":\"2018-12-17T21:08:48.527Z\"}"},
+        "response": {"req_time": 1, "error": {"code": "ECONNREFUSED", "errno": "ECONNREFUSED",
+        "syscall": "connect"}}}, {"_id": "5c145ab77a8021a8442608ed", "resource": "/items/MLA670377063",
+        "user_id": 128828674, "topic": "items", "application_id": 4607298613855815,
+        "attempts": 1, "sent": "2018-12-15T01:36:55.132Z", "received": "2018-12-15T01:36:55.062Z",
+        "request": {"url": "https://localhost:8000/auth", "headers": {"host": "localhost:8000",
+        "accept": "application/json", "content-type": "application/json", "content-length":
+        189}, "data": "{\"resource\":\"/items/MLA670377063\",\"user_id\":128828674,\"topic\":\"items\",\"application_id\":4607298613855815,\"attempts\":1,\"sent\":\"2018-12-15T01:36:55.132Z\",\"received\":\"2018-12-15T01:36:55.062Z\"}"},
+        "response": {"req_time": 2, "error": {"code": "ECONNREFUSED", "errno": "ECONNREFUSED",
+        "syscall": "connect"}}}]}'}
     headers:
       Access-Control-Allow-Headers: [Content-Type]
       Access-Control-Allow-Methods: ['PUT, GET, POST, DELETE, OPTIONS']
       Access-Control-Allow-Origin: ['*']
       Access-Control-Max-Age: ['86400']
       Connection: [keep-alive]
-      Content-Length: ['15']
+      Content-Length: ['1456']
       Content-Type: [application/json; charset=utf-8]
-      Date: ['Thu, 20 Dec 2018 18:51:25 GMT']
-      ETag: [W/"f-30c391"]
-      Via: [1.1 c3c79d4b701a0d344a641862e9711512.cloudfront.net (CloudFront)]
-      X-Amz-Cf-Id: [o20E5S17TdFrEDWK_Jvro1vpbQ7ohEu38xJWf5r-e3UPdYtacJwyKQ==]
+      Date: ['Fri, 21 Dec 2018 07:45:06 GMT']
+      ETag: [W/"bwbnsqQTwV5BbxclfWOMBQ=="]
+      Via: [1.1 af3ec872aa77de6760622a554164a644.cloudfront.net (CloudFront)]
+      X-Amz-Cf-Id: [bLVScnKfk4Xy-T51RDzwNbsmuoi0jsb7mTjkPTyaXi2PQfgVdhkgwQ==]
       X-Cache: [Miss from cloudfront]
       X-Content-Type-Options: [nosniff]
       X-Frame-Options: [DENY]
-      X-Request-Id: [4c3769af-33da-44e0-a6c7-268396f4932c]
-      X-XSS-Protection: [1; mode=block]
-    status: {code: 200, message: OK}
-- request:
-    body: grant_type=client_credentials&client_id=HIDDEN&client_secret=HIDDEN
-    headers:
-      Accept: [application/json]
-      Accept-Encoding: ['gzip, deflate']
-      Authorization: [HIDDEN]
-      Connection: [keep-alive]
-      Content-Length: ['103']
-      Content-Type: [application/x-www-form-urlencoded;charset=UTF-8]
-      User-Agent: [python-requests/2.21.0]
-    method: POST
-    uri: https://api.mercadolibre.com/oauth/token
-  response:
-    body: {string: '{"access_token": "HIDDEN", "token_type": "bearer", "expires_in":
-        21600, "scope": "marketplace_off offline_access read reporting_full wallet-payments
-        write", "user_id": "HIDDEN", "refresh_token": "HIDDEN"}'}
-    headers:
-      Access-Control-Allow-Headers: [Content-Type]
-      Access-Control-Allow-Methods: ['PUT, GET, POST, DELETE, OPTIONS']
-      Access-Control-Allow-Origin: ['*']
-      Access-Control-Max-Age: ['86400']
-      Cache-Control: [max-age=0]
-      Connection: [keep-alive]
-      Content-Type: [application/json;charset=UTF-8]
-      Date: ['Thu, 20 Dec 2018 18:51:37 GMT']
-      ETag: [942098890dab4c0688a89f7fe83dad66]
-      Strict-Transport-Security: [max-age=15724800]
-      Transfer-Encoding: [chunked]
-      Vary: ['Accept,Accept-Encoding']
-      Via: [1.1 868de9869b0df733f99027b2399ee023.cloudfront.net (CloudFront)]
-      X-Amz-Cf-Id: [G-KZBoG0jubkUIli6qwNMKCss8Jsb0NO7F0f55oVfs_6t-a0a-xNIQ==]
-      X-Cache: [Miss from cloudfront]
-      X-Content-Type-Options: [nosniff]
-      X-Frame-Options: [DENY]
-      X-Request-Id: [2530169c-2f88-4e3b-b198-4cc305ca67a5]
-      X-XSS-Protection: [1; mode=block]
-    status: {code: 200, message: OK}
-- request:
-    body: null
-    headers:
-      Accept: ['*/*']
-      Accept-Encoding: ['gzip, deflate']
-      Connection: [keep-alive]
-      User-Agent: [python-requests/2.21.0]
-    method: GET
-    uri: https://api.mercadolibre.com/myfeeds
-  response:
-    body: {string: '{"messages": []}'}
-    headers:
-      Access-Control-Allow-Headers: [Content-Type]
-      Access-Control-Allow-Methods: ['PUT, GET, POST, DELETE, OPTIONS']
-      Access-Control-Allow-Origin: ['*']
-      Access-Control-Max-Age: ['86400']
-      Connection: [keep-alive]
-      Content-Length: ['15']
-      Content-Type: [application/json; charset=utf-8]
-      Date: ['Thu, 20 Dec 2018 18:51:37 GMT']
-      ETag: [W/"f-30c391"]
-      Via: [1.1 868de9869b0df733f99027b2399ee023.cloudfront.net (CloudFront)]
-      X-Amz-Cf-Id: [Ic-G_WJHcDMLldu5FfR-vKlPrykdfnCmVq2vhjEkyhuq6SX2oqJ-Tw==]
-      X-Cache: [Miss from cloudfront]
-      X-Content-Type-Options: [nosniff]
-      X-Frame-Options: [DENY]
-      X-Request-Id: [84f0e393-dcb4-499d-87f3-9cd41b2299cd]
+      X-Request-Id: [45cac509-d858-4012-bf87-fb468a24030e]
       X-XSS-Protection: [1; mode=block]
     status: {code: 200, message: OK}
 version: 1
